Split up raster a bit more
